dag-airflow-/airflow_docker/requirements.txt
apache-airflow
evidently
pandas
numpy
psycopg2-binary
boto3
mlflow 
scikit-learn
joblib
sagemaker